reference orbitを40倍速く計算できるrust clientを使えるようにした (#50)
* Install ws

* Install tsx

* ChatGPTを頼りに作るwebsocket server/client

* 計算クライアントとして登録できるようにした

* エラー処理とかログとか追加

* blaTableは返さない

* 計算用クライアントから値を受け取れるようになった

が盛大に描画結果がぶっ壊れている

* zeroが抜けてた

* rust -> jsはJSONを経由するオーバーヘッドを減らすためバイナリで送る

* NODE_ENVを読めるようにしておく

* windows向けにcross-envしておく

* websocketつなぐ部分を別メソッドに切り出した

* 1回接続に失敗したらworker上での計算にfallbackする

* 毎回ではなくworker立ち上げ時にwebsocket接続する

* 途中で切断した場合などに対応

* README追加

* 邪魔

* なんか謎に2回wrapしてしまっていたのを修正

* workerの切り替え時にWebSocketを閉じるようにした

+++ src/worker-pool/worker-facade.ts
@@ -2,13 +2,14 @@ import {
   BatchContext,
   CalcIterationJob,
   CalcRefOrbitJob,
+  IterationIntermediateResult,
+  IterationProgress,
+  IterationResult,
   MandelbrotJob,
   MandelbrotWorkerType,
   RefOrbitProgress,
   RefOrbitResult,
-  IterationIntermediateResult,
-  IterationProgress,
-  IterationResult,
+  type RefOrbitShutdown,
 } from "@/types";
 import { refOrbitWorkerPath, workerPaths } from "@/workers";
 import { RefOrbitContext } from "@/workers/calc-ref-orbit";
@@ -95,12 +96,7 @@ export class CalcIterationWorker implements MandelbrotFacadeLike {
         case "result": {
           const { iterations, elapsed } = data;
 
-          const iterationsResult = new Uint32Array(iterations);
-
-          this.resultCallback?.(
-            { type: "result", iterations: iterationsResult, elapsed },
-            job,
-          );
+          this.resultCallback?.({ type: "result", iterations, elapsed }, job);
 
           this.worker.removeEventListener("message", f);
           this.running = false;
@@ -111,7 +107,7 @@ export class CalcIterationWorker implements MandelbrotFacadeLike {
           this.intermediateResultCallback?.(
             {
               type: "intermediateResult",
-              iterations: new Uint32Array(iterations),
+              iterations: iterations,
               resolution,
             },
             job,
@@ -285,6 +281,7 @@ export class RefOrbitWorker implements MandelbrotFacadeLike {
     Atomics.store(t, workerIdx, 0);
 
     this.worker.postMessage({
+      type: "calc-reference-orbit",
       complexCenterX,
       complexCenterY,
       pixelWidth,
@@ -304,10 +301,22 @@ export class RefOrbitWorker implements MandelbrotFacadeLike {
   };
 
   terminateAsync = () => {
-    this.worker.terminate();
+    const promise = new Promise<void>((resolve) => {
+      const handler = (ev: MessageEvent<RefOrbitShutdown>) => {
+        if (ev.data.type === "shutdown") {
+          this.worker.terminate();
+          this.running = false;
+          this.worker.removeEventListener("message", handler);
+          resolve();
+        }
+      };
 
-    this.running = false;
-    return Promise.resolve();
+      this.worker.addEventListener("message", handler);
+
+      this.worker.postMessage({ type: "request-shutdown" });
+    });
+
+    return promise;
   };
 
   cancel = ({ terminator }: BatchContext, { workerIdx }: MandelbrotJob) => {
